+++ src/main/java/com/unstoppabledomains/resolution/TickerVersion.java
@@ -1,5 +1,6 @@
 package com.unstoppabledomains.resolution;
 
+@Deprecated 
 public enum TickerVersion {
   ERC20,
   OMNI,